SVN_SILENT I hate backporting.
svn path=/branches/KDE/4.0/kdegames/konquest/; revision=772951

+++ mapscene.h
@@ -52,8 +52,7 @@ class MapScene: public QGraphicsScene
         
         void           drawBackground( QPainter * painter, const QRectF & rect );
         
-        int            itemsHorizontalOffset();
-        
+        qreal          itemsHorizontalOffset();
         qreal          getSectorSize();
     signals:
         void  planetSelected( Planet * );
